Fix init failure

+++ server/init/initializeDatabase.js
@@ -166,7 +166,9 @@ async function clearSchedule(teacher) {
       id: true,
     },
   });
-  // teacherSchedule.id may be null here
+  if(!teacherSchedule) {
+    return;
+  }
   await prisma.scheduledClass.deleteMany({
     where: {
       scheduleId: teacherSchedule.id,
